ver3/requirements.txt
# Core dependencies
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
numpy>=1.20.0
scipy>=1.7.0
librosa>=0.9.0
pyyaml>=6.0
sentencepiece>=0.1.99
einops>=0.6.0
numba>=0.56.0

# Dataset handling
datasets>=2.10.0

# LLM-related
accelerate>=0.20.0
bitsandbytes>=0.41.0  # For model quantization
optimum>=1.8.0  # For FasterTransformer
protobuf>=3.20.0

# Audio processing
audiomentations>=0.30.0
soundfile>=0.12.0

# Logging and monitoring
wandb>=0.15.0
tensorboard>=2.12.0
tqdm>=4.65.0

# Development tools
black>=23.3.0
isort>=5.12.0
pytest>=7.3.1 